+++ src/components/mypage/liked-in-brand/LikedInBrand.tsx
@@ -8,7 +8,7 @@ import {
   logoUrlNone,
   logoUrlNoneText,
   hiddenBrand, // display: none 처리를 위한 클래스 추가
-} from './liked-in-brand.css';
+} from './likedInBrand.css';
 import { useState } from 'react';
 
 type TBrand = {
