+++ src/components/mypage/size-recommendation-card/sizeRecommendationCard.css.ts
