+++ src/components/shoes-info/shoesInfo.css.ts
