+++ src/pages/mypage/viewed-history-page/viewedHistoryPage.css.ts
